nano-e-commerce/services/payment-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-reflection==1.60.0
protobuf==4.25.1
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
stripe==7.0.0
